Added vs project
- Updated to use latest IV-SDK version
