+++ IV-Presence/IV-Presence.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.9.34728.123
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "IV-Presence", "IV-Presence.vcxproj", "{0A7BF388-F78C-4AA0-94EA-FA11D7231942}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x86 = Debug|x86
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{0A7BF388-F78C-4AA0-94EA-FA11D7231942}.Debug|x86.ActiveCfg = Debug|Win32
+		{0A7BF388-F78C-4AA0-94EA-FA11D7231942}.Debug|x86.Build.0 = Debug|Win32
+		{0A7BF388-F78C-4AA0-94EA-FA11D7231942}.Release|x86.ActiveCfg = Release|Win32
+		{0A7BF388-F78C-4AA0-94EA-FA11D7231942}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {DDDE47FF-7B5E-4C6C-802C-FD3784C96E7C}
+	EndGlobalSection
+EndGlobal
